The source code tree is changed.
The source code tree is changed.
This repo have kernel and applications directory.
The kernel directory have kernel source code for Galaxy S6 now.
The x86_64 kernel code will be added to this directory.
The applications directory have source code of applications in that the fbarrier() or fdatabarrier() system call is added.