blkdev_issue_barrier() is added.
blkdev_issue_barrier() is added.
This function is same with blkdev_issue_flush() but it issue barrier command not flush command.
While blkdev_issue_flush() set the WRITE_FLUSH flag to rw_flag of dummy bio, blkdev_issue_barrier() set the WRITE_BARRIER flag to that.
It is just added but not used.
It could be used in fdatabarrier().
If the file about fdatabarrier() do not have dirty pages, fdatabarrier() can issue only barrier command as use blkdev_issue_barrier().

- Joontaek Oh